Modernisation, support for CLEF input and output
