+++ src/Seq.Apps/Apps/ISubscribeTo.cs
@@ -1,7 +1,7 @@
 ﻿namespace Seq.Apps
 {
     /// <summary>
-    /// Subscribes a reactor to a typed event.
+    /// Subscribes an app to a typed event.
     /// </summary>
     /// <typeparam name="TData">The type representing event data.</typeparam>
     public interface ISubscribeTo<TData>
